Add files via upload
0.0.1 Comcast Telecom Consumer Complaints:
Q1. Import and Analysis:
Q1. A. Import data into Python environment.
Q1. B. Provide the trend chart for the number of complaints at monthly and daily granularity levels.
Q1. C. Provide a table with the frequency of complaint types.
Q2. Which complaint types are maximum i.e., around internet, network issues, or across any other
domains.
Q3. Create a new categorical variable with value as Open and Closed. Open & Pending is to be
categorized as Open and Closed & Solved is to be categorized as Closed.
Q4. Provide state wise status of complaints in a stacked bar chart. Use the categorized variable
from Q3. Provide insights on:
Q4. A. Which state has the maximum complaints.
Q4. B. Which state has the highest percentage of unresolved complaints.
Q5. Provide the percentage of complaints resolved to date, which were received through the
Internet and customer care calls.